M3-172 Refactor : 파일 위치 변경, 환경변수 수정
