+++ .github/workflows/cd_dev.yml
@@ -30,7 +30,7 @@ jobs:
           echo "KAKAO_APP_ID=${{ secrets.KAKAO_APP_ID }}" >> .env
           echo "REDIS_HOST=${{ secrets.REDIS_HOST }}" >> .env
           echo "AWS_ACCESS_KEY=${{ secrets.AWS_ACCESS_KEY }}" >> .env
-          echo "AWS_ASECRET_KEY=${{ secrets.AWS_SECRET_KEY }}" >> .env
+          echo "AWS_SECRET_KEY=${{ secrets.AWS_SECRET_KEY }}" >> .env
 
       - name: gradlew에 실행 권한 부여
         run: chmod +x ./gradlew
