+++ src/main/java/com/m3pro/groundflip/service/UserService.java
@@ -8,7 +8,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.m3pro.groundflip.aws.S3Uploader;
+import com.m3pro.groundflip.util.S3Uploader;
 import com.m3pro.groundflip.domain.dto.user.UserInfoRequest;
 import com.m3pro.groundflip.domain.dto.user.UserInfoResponse;
 import com.m3pro.groundflip.domain.entity.User;
